changed display scale to float and gotrid of unneccessary backgrounds

+++ CrossPlatformDesktopProject/Block/BlockBrickTile.cs
@@ -8,7 +8,7 @@ namespace Sprint2
         {
             get
             {
-                return new Rectangle((int)Position.X, (int)Position.Y, Sprite.GetWidth(), (int)(Sprite.GetHeight()));
+                return new Rectangle((int)Position.X, (int)Position.Y, (int)Sprite.GetWidth(), (int)(Sprite.GetHeight()));
             }
         }
