+++ CrossPlatformDesktopProject/Interfaces/AbstractGameObject.cs
@@ -12,7 +12,7 @@ namespace Sprint2
         {
             get
             {
-                return new Rectangle((int)Position.X, (int)Position.Y, Sprite.GetWidth(), Sprite.GetHeight());
+                return new Rectangle((int)Position.X, (int)Position.Y, (int)Sprite.GetWidth(), (int)Sprite.GetHeight());
             }
         }
