+++ CrossPlatformDesktopProject/Sprite/AbstractSprite.cs
@@ -24,21 +24,21 @@ namespace Sprint2
             int column = CurrentFrame % Columns;
 
             Rectangle sourceRectangle = new Rectangle(width * column, height * row, width, height);
-            Rectangle destinationRectangle = new Rectangle((int)location.X, (int)location.Y, width * Constant.DisplayScaleX, height * Constant.DisplayScaleY);
+            Rectangle destinationRectangle = new Rectangle((int)location.X, (int)location.Y, (int)(width * Constant.DisplayScaleX), (int)(height * Constant.DisplayScaleY));
 
             spriteBatch.Draw(Texture, destinationRectangle, sourceRectangle, color);
             
         }
 
-        public virtual int GetHeight()
+        public virtual float GetHeight()
         {
-            int height = Texture.Height * Constant.DisplayScaleY / Rows;
+            float height = Texture.Height * Constant.DisplayScaleY / Rows;
             return height;
         }
 
-        public virtual int GetWidth()
+        public virtual float GetWidth()
         {
-            int width = Texture.Width * Constant.DisplayScaleX / Columns;
+            float width = Texture.Width * Constant.DisplayScaleX / Columns;
             return width;
         }
     }
