+++ CrossPlatformDesktopProject/Sprite/SpriteLinkUseItemDown.cs
@@ -22,20 +22,20 @@ namespace Sprint2
             Rectangle destinationRectangle;
 
             sourceRectangle = new Rectangle(640, 16, 16, 16);
-            destinationRectangle = new Rectangle((int)position.X, (int)position.Y, 16 * Constant.DisplayScaleX, 16 * Constant.DisplayScaleY);
+            destinationRectangle = new Rectangle((int)position.X, (int)position.Y, (int)(16 * Constant.DisplayScaleX), (int)(16 * Constant.DisplayScaleY));
 
             spriteBatch.Draw(this.Texture, destinationRectangle, sourceRectangle, color);
         }
 
-        public override int GetHeight()
+        public override float GetHeight()
         {
-            int height = Constant.LinkHeight * Constant.DisplayScaleY;
+            float height = Constant.LinkHeight * Constant.DisplayScaleY;
             return height;
         }
 
-        public override int GetWidth()
+        public override float GetWidth()
         {
-            int width = Constant.LinkWidth * Constant.DisplayScaleX;
+            float width = Constant.LinkWidth * Constant.DisplayScaleX;
             return width;
         }
     }
